+++ server/src/main/java/shinhan/mohaemoyong/server/repository/UserRepository.java
@@ -2,7 +2,7 @@ package shinhan.mohaemoyong.server.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import shinhan.mohaemoyong.server.model.entity.User;
+import shinhan.mohaemoyong.server.domain.User;
 
 import java.util.Optional;
